remove deny warnings in cli, add 'miette' feature in context crate
