+++ context/Cargo.toml
@@ -6,4 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-miette = "5.5.0"
+miette = { version = "5.5.0", optional = true }
+
+[features]
+miette = ["dep:miette"]
